moved navbar to components and changed it to do it's own auth
